pyproject.toml
[project]
name = "az-changefeed-poller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-identity>=1.16.0",
    "azure-storage-blob>=12.19.0",
    "azure-storage-blob-changefeed>=12.0.0b2",
    "boto3>=1.34.0",
    "tenacity>=8.2.0",
]

[project.scripts]
az-changefeed-poller = "main:main"
